+++ tests/translation.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 import l18n
 
 from ._base import TestCase
@@ -16,6 +18,12 @@ class TranslationTestCase(TestCase):
     def test_tz_city_translated(self):
         self.assertEqual(str(l18n.tz_cities['Canada/Newfoundland']),
                          u'Terre-Neuve')
+
+    def test_tz_fullname_translated(self):
+        self.assertEqual(
+            str(l18n.tz_fullnames['America/North_Dakota/New_Salem']),
+            u'Amérique/Dakota du Nord/New Salem'
+        )
 
     def test_territory_translated(self):
         self.assertEqual(str(l18n.territories['ES']), u'Espagne')