+++ CHANGES.rst
@@ -2,8 +2,8 @@ l18n - changes
 ==============
 
 
-v2014.10.1 (dev)
-----------------
+v2014.10.1 (17-03-2015)
+-----------------------
 
 - Birth
 - Exposes ``tz_cities``, ``tz_fullnames`` and ``territories``